Java read .csv files (No appache POI)
Reads properly, but the storing method is bugged. Works differently then C++ "Getline" (which is highly annoying. C++ does it so much easier).

So what happens is the entire thing is put into all ArrayList's, instead of the individual parameters (examples, name goes with name).

The search method works fine.